Add files via upload
First version of the control board. The basic idea was to use only a single ROM. Turns out it's possible by sorting the commands in a compact way, and for all move commands have SRC and Target addresses coded in 3 bits. But the amount of extra circuits is staggering - and does not really pay off for the "saving" of 4 or 5 ROM chips. Once I realized that I went with the other control board that is much simpler in design.